department title unique sql constraint

+++ src/main/java/com/z0976190100/departments/service/DepartmentService.java
@@ -3,41 +3,35 @@ package com.z0976190100.departments.service;
 import com.z0976190100.departments.app_constants.GeneralConstants;
 import com.z0976190100.departments.exceptions.RequestParameterValidationException;
 import com.z0976190100.departments.exceptions.ResourceNotFoundException;
+import com.z0976190100.departments.persistense.dao.Dao;
 import com.z0976190100.departments.persistense.dao.DepartmentDaoImpl;
 import com.z0976190100.departments.persistense.entity.Department;
-import com.z0976190100.departments.persistense.entity.DepartmentEntityDescription;
-import com.z0976190100.departments.persistense.entity.EntityDescription;
 
-import java.sql.SQLException;
 import java.util.List;
 
 public class DepartmentService implements GeneralConstants {
 
-    private EntityDescription description = new DepartmentEntityDescription();
-    private DepartmentDaoImpl departmentDao = new DepartmentDaoImpl();
+    private Dao<Department> dao = new DepartmentDaoImpl();
     private EmployeeService employeeService = new EmployeeService();
-    private final String table = description.getTableName();
 
 
-    // FIXME pagination- where it belongs?
-    public int getRowCount() throws SQLException {
-        String query = "SELECT COUNT(*) FROM " + table;
-        return departmentDao.getRowCount(query);
+    public int getRowCount() {
+        return ((DepartmentDaoImpl)dao).getAllRowCount();
     }
 
     public List<Department> getDepartmentsList(int offset, int limit) {
 
-        return departmentDao.getEntitiesList(offset, limit);
+        return dao.getAll(offset, limit);
     }
 
     public List<Department> getDepartmentsList() {
 
-        return departmentDao.getEntitiesList();
+        return dao.getAll();
     }
 
     public Department getDepartmentById(int id) throws ResourceNotFoundException {
 
-        Department department = departmentDao.getEntityById(id);
+        Department department = dao.getEntityById(id);
 
         if (department == null)
             throw new ResourceNotFoundException(RESOURCE_NOT_FOUND_MESSAGE + DEPARTMENT_NOT_FOUND_MESSAGE);
@@ -49,33 +43,41 @@ public class DepartmentService implements GeneralConstants {
     public void saveDepartment(String title) throws RequestParameterValidationException {
 
         // check if department is unique by title
-        isExistByTitle(title);
+        //isExistByTitle(title);
 
-        departmentDao.saveEntity(title);
+        dao.saveEntity(new Department(0, title));
 
     }
 
     public void deleteDepartment(int id) throws ResourceNotFoundException {
 
+        Department d = dao.getEntityById(id);
+
+        if (d == null) throw new ResourceNotFoundException(RESOURCE_NOT_FOUND_MESSAGE + DEPARTMENT_NOT_FOUND_MESSAGE);
+
         employeeService.deleteAllEmployees(id);
 
-        departmentDao.deleteEntity(id);
+        dao.deleteEntity(id);
 
     }
 
     public void updateDepartment(int id, String newTitle) throws ResourceNotFoundException, RequestParameterValidationException {
 
-        isExistByTitle(newTitle);
+        Department d = dao.getEntityById(id);
 
-        departmentDao.updateEntity(new Department(id, newTitle));
+        if (d == null) throw new ResourceNotFoundException(RESOURCE_NOT_FOUND_MESSAGE + DEPARTMENT_NOT_FOUND_MESSAGE);
+
+        //isExistByTitle(newTitle);
+
+        dao.updateEntity(new Department(id, newTitle));
 
     }
 
-       private void isExistByTitle(String title) throws RequestParameterValidationException {
-
-        List<Department> departmentList = departmentDao.getAllEntitiesWhere(title);
-        if (departmentList.size() != 0)
-            throw new RequestParameterValidationException(DEPARTMENT_TITLE_NOT_UNIQUE_MESSAGE);
-    }
+//       private void isExistByTitle(String title) throws RequestParameterValidationException {
+//
+//        List<Department> departmentList = dao.getAllWhere(title);
+//        if (departmentList.size() != 0)
+//            throw new RequestParameterValidationException(DEPARTMENT_TITLE_NOT_UNIQUE_MESSAGE);
+//    }
 
 }
